Refactored folder structures and installed node-sass
