+++ package.json
@@ -6,6 +6,7 @@
     "@apollo/react-hooks": "^3.1.3",
     "apollo-boost": "^0.4.4",
     "graphql": "^14.5.8",
+    "node-sass": "^4.13.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
     "react-router": "^5.1.2",
