+++ src/App.js
@@ -1,11 +1,13 @@
 import React, { useState } from "react";
 import { Route, Switch } from "react-router-dom";
-import "./App.css";
+
 import LandingPage from "./components/LandingPage";
 import Dashboard from "./components/Dashboard";
 import SignInForm from "./components/SignInForm";
 import SignUpForm from "./components/SignUpForm";
 import NavBar from "./components/NavBar";
+
+import "./global/index.scss";
 
 function App() {
   const [loggedin, setLoggedin] = useState(false);
