+++ src/components/DashboardInput/index.js
@@ -1,0 +1,3 @@
+import DashboardInput from './DashboardInput';
+
+export default DashboardInput;
