+++ src/components/ExpSignUp/index.js
@@ -1,0 +1,3 @@
+import ExpSignUp from './ExpSignUp';
+
+export default ExpSignUp;
