+++ src/components/GeneralSignUp/index.js
@@ -1,0 +1,3 @@
+import GeneralSignUp from './GeneralSignUp';
+
+export default GeneralSignUp;
