+++ src/components/NavBar/NavBar.test.js
