+++ src/components/SignInForm/SignInForm.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { useMutation } from "@apollo/react-hooks";
 import { gql } from "apollo-boost";
 // import styled from "styled-components";
-import Loading from "./Loading";
+import Loading from "../Loading";
 
 const LOGIN = gql`
   mutation Login($email: String!, $password: String!) {
