+++ src/components/SignInForm/index.js
@@ -1,0 +1,3 @@
+import SignInForm from './SignInForm';
+
+export default SignInForm;
