+++ src/components/SignUpForm/index.js
@@ -1,0 +1,3 @@
+import SignUpForm from './SignUpForm';
+
+export default SignUpForm;
