+++ src/components/UserPicture/UserPicture.js
