+++ src/components/UserPicture/index.js
@@ -1,0 +1,3 @@
+import UserPicture from './UserPicture';
+
+export default UserPicture;
