+++ src/global/App.scss
