+++ src/global/index.scss
@@ -1,0 +1,2 @@
+@import './App.scss';
+@import './test.scss';
