+++ src/global/readme.md
@@ -1,0 +1,1 @@
+Hi, I think this should work, but just make as many scss files you want in here and then import it into the index file!
